fixed a bug that was affecting multiple runs for big workloads

+++ src/bptree/memory_block_manager.h
@@ -36,7 +36,7 @@ public:
     }
 
     void reset() {
-        memset(internal_memory, 0, next_block_id * block_size);
+        memset(internal_memory, 0, (size_t) next_block_id * block_size);
         next_block_id = 0;
     }
